pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bloviate"
version = "0.2.0"
description = "Voice-fingerprinting dictation tool for whispering in noisy environments"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Bronson"}]
keywords = ["dictation", "speech-to-text", "voice", "macos", "transcription"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: MacOS X",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
  "numpy>=1.24.0",
  "sounddevice>=0.4.6",
  "scipy>=1.10.0",
  "pynput>=1.7.6",
  "torch>=2.0.0,<2.5.0",
  "torchaudio>=2.0.0,<2.5.0",
  "speechbrain==1.0.0",
  "noisereduce>=3.0.0",
  "pyqt6>=6.5.0",
  "openai-whisper>=20231117",
  "pyyaml>=6.0",
  "webrtcvad>=2.0.10",
  "websocket-client>=1.7.0",
  "keyring>=25.0.0",
]

[project.urls]
Homepage = "https://github.com/callumreid/Bloviate"
Issues = "https://github.com/callumreid/Bloviate/issues"

[project.scripts]
bloviate = "main:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = [
  "app_paths",
  "audio_capture",
  "command_vocabulary",
  "deepgram_stream",
  "history_store",
  "main",
  "model_registry",
  "noise_suppressor",
  "personal_dictionary",
  "post_processor",
  "ptt_handler",
  "secret_store",
  "settings_service",
  "torchaudio_patch",
  "transcriber",
  "ui",
  "voice_fingerprint",
  "window_manager",
]
packages = ["bloviate_resources"]

[tool.setuptools.package-data]
bloviate_resources = ["*.yaml"]
